% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomotitle.R
\name{generate_pomotitle}
\alias{generate_pomotitle}
\title{Generate a Postmodernist Title}
\usage{
generate_pomotitle(n = 1)
}
\arguments{
\item{n}{number of titles to generate.}
}
\value{
a character vector of titles.
}
\description{
Generates a character vector of postmodernist titles.
}
\details{
Why? Because we must situate erasure as a form of ontological pluralistic
framework.
}
\examples{
generate_pomotitle()

generate_pomotitle(10)
}
